Fix linking with absolute paths
Here is a case that led this change.

When we have to link against boost libraries both statically and dynamically due to dependencies requirements, we then have two library paths, one for static artifacts and the other for dynamic ones. Order matters! Let's take for example the case where the path to static libraries occurs first.

In this case, as these two paths contains library sharing same base name (boost_regex for example) then the linker that should have linked against libboost_regex.so, looking for boost_regex in first library path, tries instead to link against libboost_regex.a wrongly, even tho it is specified to link against dynamic library with -Bdynamic.

It seems that -Bdynamic doesn't enforce that the library found has to be dynamic at all. It didn't find the dynamic library, but found the static one instead in the path, it uses it, instead of finding a dynamic libdrary corresponding in other paths.